Assets added to Military Level, Tweaked with certain objects (fences) and unchecked gravity and checked is kinematic on two prefabs (server wall and machine cab.)
More about objects tweaked: I didn't want to use the same fence as the zoo level so I tweaked the metal fencing and wood fencing in the military level (the prefabs have not changed) so the player can punch the wood fencing but not the metal ones (basically its the opposite of the zoo level).

As for the prefabs, I noticed the server wall and machine cab. prefabs had gravity checked on in their rigidbodies and is kinematic unchecked. They kept falling over in the level so I unchecked gravity and checked is kinematic.